Use RSpec matcher DSL
